requirements.txt
Flask
opencv-python
numpy
vidgear
grpcio
grpcio-tools
protobuf